Move permissions to key and secret policies (#6)
I ran into problems with the other way of configuring IAM permissions. The new way seems to work and is slightly simpler to use - the user doesn't need to write the policies. A
downside is that it appears to be impossible to grant permissions to groups via resource policies, so that still must be done by extracting the `PasswordSecret` and `KmsKeyId`
outputs and writing an `AWS::IAM::Policy` or `ManagedPolicy` resource.