Clone the root before any of the copies (#24)
This is meant to save money. Clones of the root are likely to cost less.

A clone's storage costs appear to be proportional to the amount of new/changed data on it relative to the source, as long as the source lives. Once the source cluster is deleted,
the clone's full storage size is billed. If multiple clones of a source exist, they don't seem to be double-billed for the storage they share.

So to optimize cost, we try to reduce the chance that a clone will outlive its source. The longest-lived cluster is assumed to be the root, so we clone it first, if there are
available slots. Then we clone the most recent copy (which will live the longest).